requirements.txt
Flask
PyMuPDF
python-docx
spacy
nltk
transformers
openai
google-generativeai
